% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_one.R
\name{add_one}
\alias{add_one}
\title{Add one}
\usage{
add_one(x)
}
\arguments{
\item{x}{vector or number}
}
\value{
an object of the same type as \code{x} with one added to it
}
\description{
Add one
}
\examples{
j=c(2,3)
add_one(j)
}
