% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_spread.R
\name{my_spread}
\alias{my_spread}
\title{Simple spread}
\usage{
my_spread(d)
}
\arguments{
\item{d}{data frame containing columns named \code{x`` (quantitative) and }g`` (categorical)}
}
\value{
"wide" data frame with \code{g} spread, carrying along the values of \code{x}. There needs to be another variable indicating which row should contain the result.
}
\description{
Simple spread
}
\examples{
requireNamespace("tibble", quietly = TRUE)
dd=tibble::tribble(
~r, ~x, ~g,
1, 10, "a",
2, 12, "a",
3, 14, "a",
1, 15, "b",
2, 17, "b",
3, 19, "b"
)
my_spread(dd)
}
